Fix update of postgresql service
If the attribute POSTGRES_EXTERNAL_ADDRESS in secret keycloak-db-secret
is changed, then the service keycloak-postgresql has to be adopted.
If the database is not external, then the service is a simple service
with a selector and a port.
In this case we ensure that the selector and port are correct.
For an external database we create a new service to be able
to adapt to changes.
This is fixed and checked in the two new tests.